Add separate role location files
